CORE-8644 Refactor analysis wait time dialog logic into presenter
